+++ detectors/progan/artifact.py
