+++ detectors/progan/base.py
