+++ detectors/sd-v1.4/artifact.py
