+++ evaluate.py
@@ -4,146 +4,145 @@ import torch
 import numpy as np
 from tqdm import tqdm
 from PIL import Image
-from sklearn.metrics import average_precision_score
+from loguru import logger
 
-from Detectors import CospyCalibrateDetector
-from Datasets import TestDataset, EVAL_DATASET_LIST, EVAL_MODEL_LIST
-from utils import seed_torch
+from Datasets import *
+from utils import seed_torch, evaluate
 
 import warnings
 warnings.filterwarnings("ignore")
 
 
-class Detector():
-    def __init__(self, args):
-        super(Detector, self).__init__()
+# Load pre-trained models for evaluation
+class Detector:
+    def __init__(self,
+                 device: str,
+                 mode: str = "fusion",
+                 train_dataset: str = "sd-v1.4",
+                 ckpt: str = "ckpt",
+                 batch_size: int = 32):
 
         # Device
-        self.device = args.device
+        self.device = device
+        self.mode = mode
+        self.train_dataset = train_dataset
+        self.ckpt = ckpt
+        self.batch_size = batch_size
 
-        # Initialize the detector
-        self.model = CospyCalibrateDetector(
-            semantic_weights_path=args.semantic_weights_path,
-            artifact_weights_path=args.artifact_weights_path)
+        # Dynamically import the detector module from either "progan" or "sd-v1.4"
+        detector_module = importlib.import_module(f"detectors.{train_dataset}")
 
-        # Load the pre-trained weights
-        self.model.load_weights(args.classifier_weights_path)
-        self.model.eval()
-
-        # Put the model on the device
+        # Get the detector and load weights based on mode
+        if self.mode == "fusion":
+            semantic_weights_path = os.path.join(self.ckpt, self.train_dataset, "semantic", "best_model.pth")
+            artifact_weights_path = os.path.join(self.ckpt, self.train_dataset, "artifact", "best_model.pth")
+            fusion_weights_path = os.path.join(self.ckpt, self.train_dataset, "fusion", "best_model.pth")
+            if not os.path.exists(semantic_weights_path) or not os.path.exists(artifact_weights_path) or not os.path.exists(fusion_weights_path):
+                raise ValueError("Semantic, Artifact or Fusion weights path does not exist for fusion mode")
+            CoSpyFusionDetector = getattr(detector_module, "CoSpyFusionDetector")
+            self.model = CoSpyFusionDetector(
+                semantic_weights_path=semantic_weights_path,
+                artifact_weights_path=artifact_weights_path)
+            self.model.load_weights(fusion_weights_path)
+        elif self.mode == "end2end":
+            End2EndDetector = getattr(detector_module, "End2EndDetector")
+            self.model = End2EndDetector()
+            end2end_weights_path = os.path.join(self.ckpt, self.train_dataset, "end2end", "best_model.pth")
+            if not os.path.exists(end2end_weights_path):
+                raise ValueError("End2End weights path does not exist for end2end mode")
+            self.model.load_weights(end2end_weights_path)
+        else:
+            raise ValueError(f"Unknown mode: {self.mode}")
+        
+        # Put the model on the device and set to eval
         self.model.to(self.device)
-
-    # Prediction function
-    def predict(self, inputs):
-        inputs = inputs.to(self.device)
-        outputs = self.model(inputs)
-        prediction = outputs.sigmoid().flatten().tolist()
-        return prediction
-
-
-def evaluate(y_pred, y_true):
-    ap = average_precision_score(y_true, y_pred)
-    accuracy = ((np.array(y_pred) > 0.5) == y_true).mean()
-    return ap, accuracy
-
-
-def test(args):
-    # Initialize the detector
-    detector = Detector(args)
-
-    # Set the saving directory
-    if not os.path.exists(args.save_dir):
-        os.makedirs(args.save_dir)
-    save_result_path = os.path.join(args.save_dir, "result.json")
-    save_output_path = os.path.join(args.save_dir, "output.json")
-
-    # Begin the evaluation
-    result_all = {}
-    output_all = {}
-    for dataset_name in EVAL_DATASET_LIST:
-        result_all[dataset_name] = {}
-        output_all[dataset_name] = {}
-        for model_name in EVAL_MODEL_LIST:
-            test_dataset = TestDataset(dataset=dataset_name, model=model_name, root_path=args.testset_dirpath, transform=detector.model.test_transform)
-            test_loader = torch.utils.data.DataLoader(test_dataset,
-                                                      batch_size=args.batch_size,
-                                                      shuffle=False,
-                                                      num_workers=4,
-                                                      pin_memory=True)
-
-            # Evaluate the model
-            y_pred, y_true = [], []
-            for (images, labels) in tqdm(test_loader, desc=f"Evaluating {dataset_name} {model_name}"):
-                y_pred.extend(detector.predict(images))
-                y_true.extend(labels.tolist())
-
-            ap, accuracy = evaluate(y_pred, y_true)
-            print(f"Evaluate on {dataset_name} {model_name} | Size {len(y_true)} | AP {ap*100:.2f}% | Accuracy {accuracy*100:.2f}%")
-
-            result_all[dataset_name][model_name] = {"size": len(y_true), "AP": ap, "Accuracy": accuracy}
-            output_all[dataset_name][model_name] = {"y_pred": y_pred, "y_true": y_true}
-
-    # Save the results
-    with open(save_result_path, "w") as f:
-        json.dump(result_all, f, indent=4)
-
-    with open(save_output_path, "w") as f:
-        json.dump(output_all, f, indent=4)
-
-
-def scan(args):
-    # Initialize the detector
-    detector = Detector(args)
-
-    # Define the pre-processing function
-    test_transform = detector.model.test_transform
-
-    # Load the image
-    image_filepath = input("Please enter the image filepath for scanning: ")
-    if not os.path.exists(image_filepath):
-        print(f"Image file not found: {image_filepath}")
-        image_filepath = input("Please enter the image filepath for scanning: ")
-
-    image = Image.open(image_filepath).convert("RGB")
-    image = test_transform(image)
-    image = image.unsqueeze(0)
-    image = image.to(args.device)
+        self.model.eval()
     
-    # Make the prediction
-    prediction = detector.predict(image)[0]
+    def evaluate_benchmark(self):
+        # Select the appropriate test dataset and evaluation lists
+        if self.train_dataset == "progan":
+            benchmark_name = "AIGCDetectionBenchMark"
+            TestDataset = ProGANTestDataset
+            eval_dataset_list = AIGCDetectionBenchMark_DATASET_LIST
+            eval_model_list = AIGCDetectionBenchMark_MODEL_LIST
+        elif self.train_dataset == "sd-v1.4":
+            benchmark_name = "Co-Spy-Bench"
+            TestDataset = CoSpyBenchTestDataset
+            eval_dataset_list = CoSpyBench_DATASET_LIST
+            eval_model_list = CoSpyBench_MODEL_LIST
+        else:
+            raise ValueError(f"Unknown train dataset: {self.train_dataset}")
 
-    if prediction > 0.5:
-        print(f"CO-SPY Prediction: {prediction:.3f} - AI-Generated")
-    else:
-        print(f"CO-SPY Prediction: {prediction:.3f} - Real")
+        # Set the saving directory
+        save_dir = os.path.join(self.ckpt, self.train_dataset, self.mode, f"eval_{benchmark_name}")
+        if not os.path.exists(save_dir):
+            os.makedirs(save_dir)
+        
+        # Setup logger
+        log_path = f"{save_dir}/evaluation.log"
+        if os.path.exists(log_path):
+            os.remove(log_path)
+        
+        logger_id = logger.add(
+            log_path,
+            format="{time:MM-DD at HH:mm:ss} | {level} | {module}:{line} | {message}",
+            level="DEBUG",
+        )
 
+        # Save raw model prediction
+        save_output_path = os.path.join(save_dir, "output.json")
+        # Save summarized evaluation result
+        save_result_path = os.path.join(save_dir, "result.json")
 
-if __name__ == "__main__":
-    import argparse
+        # Begin the evaluation
+        result_all = {}
+        output_all = {}
+        for dataset_name in eval_dataset_list:
+            result_all[dataset_name] = {}
+            output_all[dataset_name] = {}
+            for model_name in eval_model_list:
+                test_dataset = TestDataset(dataset=dataset_name, model=model_name, transform=self.model.test_transform)
+                test_loader = torch.utils.data.DataLoader(test_dataset,
+                                                          batch_size=args.batch_size,
+                                                          shuffle=False,
+                                                          num_workers=4,
+                                                          pin_memory=True)
 
-    parser = argparse.ArgumentParser("Deep Fake Detection")
-    parser.add_argument("--gpu", type=int, default=0, help="GPU ID")
-    parser.add_argument("--phase", type=str, default="scan", choices=["scan", "test"], help="Phase of the experiment")
-    parser.add_argument("--semantic_weights_path", type=str, default="pretrained/semantic_weights.pth", help="Semantic weights path")
-    parser.add_argument("--artifact_weights_path", type=str, default="pretrained/artifact_weights.pth", help="Artifact weights path")
-    parser.add_argument("--classifier_weights_path", type=str, default="pretrained/classifier_weights.pth", help="Classifier weights path")
-    parser.add_argument("--testset_dirpath", type=str, default="data/test", help="Testset directory")
-    parser.add_argument("--save_dir", type=str, default="test_results", help="Save directory")
-    parser.add_argument("--batch_size", type=int, default=32, help="Batch size")
-    parser.add_argument("--seed", type=int, default=1024, help="Random seed")
+                # Evaluate the model
+                y_pred, y_true = [], []
+                for (images, labels) in tqdm(test_loader, desc=f"Evaluating {benchmark_name} - {dataset_name} - {model_name}"):
+                    y_pred.extend(detector.predict(images))
+                    y_true.extend(labels.tolist())
 
-    args = parser.parse_args()
+                ap, accuracy = evaluate(y_pred, y_true)
+                logger.info(f"Evaluate on {benchmark_name} - {dataset_name} - {model_name} | Size {len(y_true)} | AP {ap*100:.2f}% | Accuracy {accuracy*100:.2f}%")
 
-    # Set the random seed
-    seed_torch(args.seed)
+                result_all[dataset_name][model_name] = {"size": len(y_true), "AP": ap, "Accuracy": accuracy}
+                output_all[dataset_name][model_name] = {"y_pred": y_pred, "y_true": y_true}
 
-    # Set the GPU ID
-    args.device = f"cuda:{args.gpu}" if torch.cuda.is_available() else "cpu"
+        # Save the results
+        with open(save_result_path, "w") as f:
+            json.dump(result_all, f, indent=4)
 
-    # Begin the experiment
-    if args.phase == "scan":
-        scan(args)
-    elif args.phase == "test":
-        test(args)
-    else:
-        raise ValueError("Unknown phase")
+        with open(save_output_path, "w") as f:
+            json.dump(output_all, f, indent=4)
+
+    def scan(self):
+        # Load the image
+        image_filepath = input("Please enter the image filepath for scanning: ")
+        if not os.path.exists(image_filepath):
+            print(f"Image file not found: {image_filepath}")
+            image_filepath = input("Please enter the image filepath for scanning: ")
+
+        image = Image.open(image_filepath).convert("RGB")
+        image = self.model.test_transform(image)
+        image = image.unsqueeze(0)
+        image = image.to(self.device)
+
+        # Make the prediction
+        prediction = self.predict(image)[0]
+
+        if prediction > 0.5:
+            print(f"Co-Spy Prediction: {prediction:.3f} - AI-Generated")
+        else:
+            print(f"Co-Spy Prediction: {prediction:.3f} - Real")
