+++ train.py
@@ -1,69 +1,98 @@
 import os
 import time
-import json
 import torch
-import numpy as np
-from tqdm import tqdm
+import importlib
 from loguru import logger
-from sklearn.metrics import average_precision_score
 
-from utils import seed_torch
-from Detectors import CospyDetector, LabelSmoothingBCEWithLogits
-from Datasets import TrainDataset, TestDataset, EVAL_DATASET_LIST, EVAL_MODEL_LIST
+from utils import seed_torch, evaluate
+from datasets import TrainDataset
 
 import warnings
 warnings.filterwarnings("ignore")
 
 
-class Detector():
-    def __init__(self, args):
-        super(Detector, self).__init__()
+class Trainer:
+    def __init__(self,
+                 mode: str,
+                 device: str,
+                 branch: str,
+                 train_dataset: str,
+                 label_smooth: bool = False,
+                 ckpt: str = "ckpt",
+                 epochs: int = 20,
+                 batch_size: int = 32):
 
-        # Device
-        self.device = args.device
+        self.device = device
+        self.mode = mode
+        self.branch = branch
+        self.train_dataset = train_dataset
+        self.semantic_weights_path = semantic_weights_path
+        self.artifact_weights_path = artifact_weights_path
+        self.label_smooth = label_smooth
+        self.ckpt = ckpt
+        self.epochs = epochs
+        self.batch_size = batch_size
 
-        # Get the detector
-        self.model = CospyDetector()
+        # Dynamically import the detector module from either "progan" or "sd-v1.4"
+        detector_module = importlib.import_module(f"detectors.{train_dataset}")
 
-        # Put the model on the device
+        # Get the detector based on mode
+        if self.mode == "branch":
+            ArtifactDetector = getattr(detector_module, "ArtifactDetector")
+            SemanticDetector = getattr(detector_module, "SemanticDetector")
+            if self.branch == "artifact":
+                self.model = ArtifactDetector()
+            elif self.branch == "semantic":
+                self.model = SemanticDetector()
+            else:
+                raise ValueError(f"Unknown detector: {self.branch}")
+        elif self.mode == "fusion":
+            semantic_weights_path = os.path.join(self.ckpt, self.train_dataset, "semantic", "best_model.pth")
+            artifact_weights_path = os.path.join(self.ckpt, self.train_dataset, "artifact", "best_model.pth")
+            if not os.path.exists(semantic_weights_path) or not os.path.exists(artifact_weights_path):
+                raise ValueError("Semantic or Artifact weights path does not exist for fusion mode")
+            CoSpyFusionDetector = getattr(detector_module, "CoSpyFusionDetector")
+            self.model = CoSpyFusionDetector(
+                semantic_weights_path=semantic_weights_path,
+                artifact_weights_path=artifact_weights_path)
+        elif self.mode == "end2end":
+            End2EndDetector = getattr(detector_module, "End2EndDetector")
+            self.model = End2EndDetector()
+        else:
+            raise ValueError(f"Unknown mode: {self.mode}")
+
         self.model.to(self.device)
 
         # Initialize the fc layer
-        torch.nn.init.normal_(self.model.sem.fc.weight.data, 0.0, 0.02)
-        torch.nn.init.normal_(self.model.art.fc.weight.data, 0.0, 0.02)
         torch.nn.init.normal_(self.model.fc.weight.data, 0.0, 0.02)
+        if self.mode == "end2end":
+            torch.nn.init.normal_(self.model.sem.fc.weight.data, 0.0, 0.02)
+            torch.nn.init.normal_(self.model.art.fc.weight.data, 0.0, 0.02)
 
         # Optimizer
-        _lr = 1e-4
         _beta1 = 0.9
         _weight_decay = 0.0
-        params = []
-        for name, param in self.model.named_parameters():
-            if param.requires_grad:
-                params.append(param)
-        print(f"Trainable parameters: {len(params)}")
+        params = [p for p in self.model.parameters() if p.requires_grad]
+        logger.info(f"Trainable parameters: {len(params)}")
 
-        self.optimizer = torch.optim.AdamW(params, lr=_lr, betas=(_beta1, 0.999), weight_decay=_weight_decay)
+        self._lr = 1e-4 if self.mode != "fusion" else 1e-1
+        self.optimizer = torch.optim.AdamW(params, lr=self._lr, betas=(_beta1, 0.999), weight_decay=_weight_decay)
 
         # Loss function
-        if args.no_label_smooth:
-            self.criterion = torch.nn.BCEWithLogitsLoss()
-        else:
+        if self.label_smooth:
             self.criterion = LabelSmoothingBCEWithLogits(smoothing=0.1)
+        else:
+            self.criterion = torch.nn.BCEWithLogitsLoss()
 
         # Scheduler
         self.delr_freq = 10
 
-    # Training function for the detector
     def train_step(self, batch_data):
-        # Decompose the batch data
         inputs, labels = batch_data
         inputs, labels = inputs.to(self.device), labels.to(self.device)
 
         self.optimizer.zero_grad()
-
         outputs = self.model(inputs)
-
         loss = self.criterion(outputs, labels.unsqueeze(1).float())
         loss.backward()
         self.optimizer.step()
@@ -73,199 +102,112 @@ class Detector():
         y_true = labels.tolist()
         return eval_loss, y_pred, y_true
 
-    # Schedule the training
-    # Early stopping / learning rate adjustment
     def scheduler(self, status_dict):
         epoch = status_dict["epoch"]
         if epoch % self.delr_freq == 0 and epoch != 0:
             for param_group in self.optimizer.param_groups:
                 param_group["lr"] *= 0.9
-            self.lr = param_group["lr"]
-        return True
-    
-    # Prediction function
-    def predict(self, inputs):
-        inputs = inputs.to(self.device)
-        outputs = self.model(inputs)
-        prediction = outputs.sigmoid().flatten().tolist()
-        return prediction
+            self._lr = param_group["lr"]
 
-
-def evaluate(y_pred, y_true):
-    ap = average_precision_score(y_true, y_pred)
-    accuracy = ((np.array(y_pred) > 0.5) == y_true).mean()
-    return ap, accuracy
-
-
-def train(args):
-    # Get the detector
-    detector = Detector(args)
-
-    # Load the dataset
-    train_dataset = TrainDataset(data_path=args.trainset_dirpath,
-                                 split="train",
-                                 transform=detector.model.train_transform)
-    train_loader  = torch.utils.data.DataLoader(train_dataset,
-                                                batch_size=args.batch_size,
-                                                shuffle=True,
-                                                num_workers=4,
-                                                pin_memory=True)
-
-    test_dataset  = TrainDataset(data_path=args.trainset_dirpath,
-                                 split="val",
-                                 transform=detector.model.test_transform)
-    test_loader   = torch.utils.data.DataLoader(test_dataset,
-                                                batch_size=args.batch_size,
-                                                shuffle=False,
-                                                num_workers=4,
-                                                pin_memory=True)
-
-    logger.info(f"Train size {len(train_dataset)} | Test size {len(test_dataset)}")
-
-    # Set the saving directory
-    model_dir = os.path.join(args.ckpt, "cospy")
-    if not os.path.exists(model_dir):
-        os.makedirs(model_dir)
-    log_path = f"{model_dir}/training.log"
-    if os.path.exists(log_path):
-        os.remove(log_path)
-
-    logger_id = logger.add(
-        log_path,
-        format="{time:MM-DD at HH:mm:ss} | {level} | {module}:{line} | {message}",
-        level="DEBUG",
-    )
-
-    # Train the detector
-    best_acc = 0
-    for epoch in range(args.epochs):
-        # Set the model to training mode
-        detector.model.train()
-        time_start = time.time()
-        for step_id, batch_data in enumerate(train_loader):
-            eval_loss, y_pred, y_true = detector.train_step(batch_data)
-            ap, accuracy = evaluate(y_pred, y_true)
-
-            # Log the training information
-            if (step_id + 1) % 100 == 0:
-                time_end = time.time()
-                logger.info(f"Epoch {epoch} | Batch {step_id + 1}/{len(train_loader)} | Loss {eval_loss:.4f} | AP {ap*100:.2f}% | Accuracy {accuracy*100:.2f}% | Time {time_end-time_start:.2f}s")
-                time_start = time.time()
+    def train(self):
+        # Determine data split and transform based on mode
+        if self.mode == "fusion":
+            train_split, val_split = "val", "val"
+            train_transform = self.model.test_transform
+            test_transform = self.model.test_transform
+        else:  # branch or end2end mode
+            train_split, val_split = "train", "val"
+            train_transform = self.model.train_transform
+            test_transform = self.model.test_transform
         
-        # Evaluate the model
-        detector.model.eval()
-        y_pred, y_true = [], []
-        for (images, labels) in test_loader:
-            y_pred.extend(detector.predict(images))
-            y_true.extend(labels.tolist())
+        # Determine save directory
+        if self.mode == "branch":
+            subdir = self.branch
+        else:
+            subdir = self.mode
+        # Set the saving directory
+        model_dir = os.path.join(self.ckpt, self.train_dataset, subdir)
+        if not os.path.exists(model_dir):
+            os.makedirs(model_dir)
+        
+        # Setup logger
+        log_path = f"{model_dir}/training.log"
+        if os.path.exists(log_path):
+            os.remove(log_path)
+        
+        logger_id = logger.add(
+            log_path,
+            format="{time:MM-DD at HH:mm:ss} | {level} | {module}:{line} | {message}",
+            level="DEBUG",
+        )
 
-        ap, accuracy = evaluate(y_pred, y_true)
-        logger.info(f"Epoch {epoch} | Test AP {ap*100:.2f}% | Test Accuracy {accuracy*100:.2f}%")
+        # Add JPEG compression for sd-v1.4 dataset
+        self.add_jpeg = True if self.train_dataset == "sd-v1.4" else False
 
-        # Schedule the training
-        status_dict = {"epoch": epoch, "AP": ap, "Accuracy": accuracy}
-        proceed = detector.scheduler(status_dict)
-        if not proceed:
-            logger.info("Early stopping")
-            break
+        # Load the training and validation dataset
+        train_dataset = TrainDataset(train_dataset=self.train_dataset,
+                                     split=train_split,
+                                     add_jpeg=self.add_jpeg,
+                                     transform=train_transform)
+        train_loader = torch.utils.data.DataLoader(train_dataset,
+                                                   batch_size=self.batch_size,
+                                                   shuffle=True,
+                                                   num_workers=4,
+                                                   pin_memory=True)
+        val_dataset = TrainDataset(train_dataset=self.train_dataset,
+                                   split=val_split,
+                                   add_jpeg=self.add_jpeg,
+                                   transform=test_transform)
+        val_loader = torch.utils.data.DataLoader(val_dataset,
+                                                 batch_size=self.batch_size,
+                                                 shuffle=False,
+                                                 num_workers=4,
+                                                 pin_memory=True)
 
-        # Save the model
-        if accuracy >= best_acc:
-            best_acc = accuracy
-            detector.model.save_weights(f"{model_dir}/best_model.pth")
-            logger.info(f"Best model saved with accuracy {best_acc.mean()*100:.2f}%")
+        logger.info(f"Train size {len(train_dataset)} | Val size {len(val_dataset)}")
 
-        if epoch % 5 == 0:
-            detector.model.save_weights(f"{model_dir}/epoch_{epoch}.pth")
-            logger.info(f"Model saved at epoch {epoch}")
+        # Train the detector
+        best_acc = 0
+        for epoch in range(self.epochs):
+            self.model.train()
+            time_start = time.time()
+            for step_id, batch_data in enumerate(train_loader):
+                eval_loss, y_pred, y_true = self.train_step(batch_data)
+                ap, accuracy = evaluate(y_pred, y_true)
 
-    # Save the final model
-    detector.model.save_weights(f"{model_dir}/final_model.pth")
-    logger.info("Final model saved")
-
-    # Remove the logger
-    logger.remove(logger_id)
-
-
-def test(args):
-    # Initialize the detector
-    detector = Detector(args)
-
-    # Load the [best/final] model
-    weights_path = os.path.join(args.ckpt, "cospy", "best_model.pth")
-
-    detector.model.load_weights(weights_path)
-    detector.model.to(args.device)
-    detector.model.eval()
-
-    # Set the pre-processing function
-    test_transform = detector.model.test_transform
-
-    # Set the saving directory
-    save_dir = os.path.join(args.ckpt, "cospy")
-    save_result_path = os.path.join(save_dir, "result.json")
-    save_output_path = os.path.join(save_dir, "output.json")
-
-    # Begin the evaluation
-    result_all = {}
-    output_all = {}
-    for dataset_name in EVAL_DATASET_LIST:
-        result_all[dataset_name] = {}
-        output_all[dataset_name] = {}
-        for model_name in EVAL_MODEL_LIST:
-            test_dataset = TestDataset(dataset=dataset_name, model=model_name, root_path=args.testset_dirpath, transform=test_transform)
-            test_loader = torch.utils.data.DataLoader(test_dataset,
-                                                      batch_size=args.batch_size,
-                                                      shuffle=False,
-                                                      num_workers=4,
-                                                      pin_memory=True)
+                if (step_id + 1) % 100 == 0:
+                    time_end = time.time()
+                    logger.info(f"Epoch {epoch} | Batch {step_id + 1}/{len(train_loader)} | Loss {eval_loss:.4f} | AP {ap*100:.2f}% | Accuracy {accuracy*100:.2f}% | Time {time_end-time_start:.2f}s")
+                    time_start = time.time()
 
             # Evaluate the model
+            self.model.eval()
             y_pred, y_true = [], []
-            for (images, labels) in tqdm(test_loader, desc=f"Evaluating {dataset_name} {model_name}"):
-                y_pred.extend(detector.predict(images))
+            for (images, labels) in val_loader:
+                y_pred.extend(self.model.predict(images))
                 y_true.extend(labels.tolist())
 
             ap, accuracy = evaluate(y_pred, y_true)
-            print(f"Evaluate on {dataset_name} {model_name} | Size {len(y_true)} | AP {ap*100:.2f}% | Accuracy {accuracy*100:.2f}%")
+            eval_type = "Test" if self.mode == "branch" else "Total"
+            logger.info(f"Epoch {epoch} | {eval_type} AP {ap*100:.2f}% | {eval_type} Accuracy {accuracy*100:.2f}%")
 
-            result_all[dataset_name][model_name] = {"size": len(y_true), "AP": ap, "Accuracy": accuracy}
-            output_all[dataset_name][model_name] = {"y_pred": y_pred, "y_true": y_true}
+            # Schedule the training
+            status_dict = {"epoch": epoch, "AP": ap, "Accuracy": accuracy}
+            self.scheduler(status_dict)
 
-    # Save the results
-    with open(save_result_path, "w") as f:
-        json.dump(result_all, f, indent=4)
+            # Save the model
+            if accuracy >= best_acc:
+                best_acc = accuracy
+                self.model.save_weights(f"{model_dir}/best_model.pth")
+                logger.info(f"Best model saved with accuracy {best_acc*100:.2f}%")
 
-    with open(save_output_path, "w") as f:
-        json.dump(output_all, f, indent=4)
+            if epoch % 5 == 0:
+                self.model.save_weights(f"{model_dir}/epoch_{epoch}.pth")
+                logger.info(f"Model saved at epoch {epoch}")
 
+        # Save the final model
+        self.model.save_weights(f"{model_dir}/final_model.pth")
+        logger.info("Final model saved")
 
-if __name__ == "__main__":
-    import argparse
-
-    parser = argparse.ArgumentParser("Deep Fake Detection")
-    parser.add_argument("--gpu", type=int, default=0, help="GPU ID")
-    parser.add_argument("--phase", type=str, default="test", choices=["train", "test"], help="Phase of the experiment")
-    parser.add_argument("--no_label_smooth", action="store_true", help="Whether to use label smoothing")
-    parser.add_argument("--trainset_dirpath", type=str, default="data/train", help="Trainset directory")
-    parser.add_argument("--testset_dirpath", type=str, default="data/test", help="Testset directory")
-    parser.add_argument("--ckpt", type=str, default="ckpt", help="Checkpoint directory")
-    parser.add_argument("--epochs", type=int, default=10, help="Number of epochs")
-    parser.add_argument("--batch_size", type=int, default=32, help="Batch size")
-    parser.add_argument("--seed", type=int, default=1024, help="Random seed")
-
-    args = parser.parse_args()
-
-    # Set the random seed
-    seed_torch(args.seed)
-
-    # Set the GPU ID
-    args.device = f"cuda:{args.gpu}" if torch.cuda.is_available() else "cpu"
-
-    # Begin the experiment
-    if args.phase == "train":
-        train(args)
-    elif args.phase == "test":
-        test(args)
-    else:
-        raise ValueError("Unknown phase")
+        # Remove the logger
+        logger.remove(logger_id)
